Add middleware in every endpoint

+++ src/routes/api/chat/message/delete/[messageId]/+server.js
@@ -1,17 +1,21 @@
 import { json } from '@sveltejs/kit';
 import Message from '../../../../../../lib/modals/message.js';
 
-export async function DELETE({ params }) {
+export async function DELETE({ params, locals }) {
     try {
+      if (!locals.userId) {
+        return json({ error: "Unauthorized" }, { status: 401 });
+      }
+
       const { messageId } = params;
       const messageGet = await Message.deleteOne({ _id: messageId });
       if (messageGet) {
-        return json({ status: 200, message: "Message was deleted successfully" });
+        return json({ message: "Message was deleted successfully" }, { status: 200 });
       } else {
-        return json({ status: 404, error: "Message was not found" });
+        return json({ error: "Message was not found" }, { status: 404 });
       }
     } catch (error) {
       console.log(error)
-      return json({ status: 500, error: "Internal server error" });
+      return json({ error: "Internal server error" }, { status: 500 });
     }
 }
